CU-86c029q15 - Update Windows Installation Scripts

+++ JeMPI_Apps/JeMPI_LibAPI/src/main/java/org/jembi/jempi/libapi/BackEnd.java
@@ -20,7 +20,6 @@ import org.jembi.jempi.shared.models.ConfigurationModel.Configuration;
 import org.jembi.jempi.shared.models.dashboard.NotificationStats;
 import org.jembi.jempi.shared.models.dashboard.SQLDashboardData;
 import org.jembi.jempi.shared.utils.AppUtils;
-
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
@@ -469,7 +468,14 @@ public final class BackEnd extends AbstractBehavior<BackEnd.Event> {
          if (request.uploadConfig != null) {
             appendUploadConfigToFile(request.uploadConfig, file);
          }
-         Files.move(file.toPath(), Paths.get("/app/csv/" + file.getName()));
+         Path ubuntuFilePath = new File(String.format("/app/csv")).toPath();
+         if (Files.exists(ubuntuFilePath)) {
+            Files.move(file.toPath(), Paths.get("/app/csv/" + file.getName()));
+         } else {
+            final String configDir = System.getenv("SYSTEM_CSV_DIR");
+            Path filePath = Paths.get(configDir, file.getName());
+            Files.move(file.toPath(), filePath);
+         }
       } catch (NoSuchFileException e) {
          LOGGER.error("No such file");
       } catch (SecurityException | IOException e) {
@@ -510,8 +516,17 @@ public final class BackEnd extends AbstractBehavior<BackEnd.Event> {
 
    private Behavior<Event> getFieldsConfigurationHandler(final GetFieldsConfigurationRequest request) {
       final var separator = FileSystems.getDefault().getSeparator();
-      final var filePath =
-            new File("%sapp%sconf_system%s%s".formatted(separator, separator, separator, fieldsConfigurationFileName)).toPath();
+      final String configDir = System.getenv("SYSTEM_CONFIG_DIRS");
+      Path filePath = Paths.get(""); // Start with an empty path
+        // Create ubuntuFilePath
+      Path ubuntuFilePath = new File(String.format("%sapp%sconf_system%s%s", separator, separator, separator, fieldsConfigurationFileName)).toPath();
+      // Check if ubuntuFilePath exists
+      if (Files.exists(ubuntuFilePath)) {
+         filePath = ubuntuFilePath;
+      } else {
+         // If ubuntuFilePath does not exist, assign the alternative path
+         filePath = Paths.get(configDir, "config-api.json");
+      }
       try {
          String configFileContent = new String(Files.readAllBytes(filePath), StandardCharsets.UTF_8);
          FieldsConfiguration fieldsConfiguration = AppUtils.OBJECT_MAPPER.readValue(configFileContent, FieldsConfiguration.class);
